Fixes:
- Refactor the way Player::config_game is initialized, so that the values set by indirectly adressing config params don't get overwritten.
- Replace all occurrences of ':' for INI parameters with '.'
- Game config parameters were passed by copy instead of reference in some occassions
- Also implement EXPlus modifier for RPG2k3
- Fix GCC warnings

+++ src/scene_battle_rpg2k.cpp
@@ -882,7 +882,7 @@ Scene_Battle_Rpg2k::SceneActionReturn Scene_Battle_Rpg2k::ProcessSceneActionVict
 
 		pm.PushPageEnd();
 
-		for (int i = 0; i < ally_battlers.size(); ++i) {
+		for (int i = 0; i < static_cast<int>(ally_battlers.size()); ++i) {
 			Game_Actor* actor = static_cast<Game_Actor*>(ally_battlers[i]);
 			int exp_gain = exp;
 			RuntimePatches::EXPlus::ModifyExpGain(*actor, exp_gain);
